requirements.txt
fastapi
uvicorn
requests
python-dotenv
groq
streamlit